+++ packages/components/src/autocomplete/index.tsx
@@ -444,6 +444,7 @@ export function useAutocompleteProps( options: UseAutocompleteProps ) {
 			function _onKeyDown( event: KeyboardEvent ) {
 				onKeyDownRef.current?.( event );
 			}
+			// To do: comment.
 			element.ownerDocument.addEventListener( 'keydown', _onKeyDown );
 			element.addEventListener( 'keydown', _onKeyDown );
 			return () => {